modified:   addLike.js 	modified:   api.js 	modified:   rendering.js
